requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
Markdown==3.7
numpy==2.2.1
openai==1.59.3
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
